requirements.txt
pandas
numpy
pyyaml
matplotlib
streamlit
